backup/requirements_original.txt
fastapi==0.100.0
uvicorn==0.23.1
duckdb==0.8.1
redis==4.6.0
pandas==2.0.3
requests==2.31.0
yfinance==0.2.65
openpyxl==3.1.2
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.0
lxml>=4.9.0
html5lib>=1.1
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
aiofiles>=0.8.0
